grammar/transform: Correct past active participle for type three
The previous implementation was not duplicating the third last letter.
This is important in the case of verbs ending with -sta or -stä, which
will add a -sut or -syt ending, coming from the initial s. Other type
three verbs worked because their third last and second last letter at
the same, -sta and -stä are the only cases where this does not hold true.
Duplicate the third last letter.